Fixed getDataNames
