+++ package.json
@@ -54,9 +54,7 @@
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-react": "^7.11.1",
-    "typescript": "^3.2.1"
-  },
-  "dependencies": {
+    "typescript": "^3.2.1",
     "prettier": "^1.15.3"
   }
 }
