+++ src/aliases/index.ts
@@ -1,0 +1,2 @@
+export { default } from './makeAliases';
+export * from './types';
