+++ src/aliases/types.ts
@@ -1,3 +1,3 @@
-import { PartialBreakpoints } from './components/ThemeProvider/ThemeProvider.types';
+import { PartialBreakpoints } from '../components/ThemeProvider';
 
 export type MakeAliases = (breakpoints: PartialBreakpoints | undefined) => PartialBreakpoints;
