+++ src/components/BaseCSS/types.ts
