+++ src/index.ts
@@ -5,10 +5,9 @@ export { default as Col } from './components/Col';
 export { default as BaseCSS } from './components/BaseCSS';
 export { default as GridThemeProvider } from './components/ThemeProvider';
 
-export * from './components/BaseCSS/BaseCSS.types';
-export * from './components/Col/Col.types';
-export * from './components/Container/Container.types';
-export * from './components/Row/Row.types';
-export * from './components/ThemeProvider/ThemeProvider.types';
-export * from './aliases.types';
-export * from './media.types';
+export * from './components/BaseCSS/types';
+export * from './components/Col/types';
+export * from './components/Container/types';
+export * from './components/Row/types';
+export * from './components/ThemeProvider/types';
+export * from './media/types';
