+++ src/media/index.ts
@@ -1,0 +1,2 @@
+export { default } from './media';
+export * from './types';
